Update README.md
This Random Forest classifier model was developed to detect security threats in a Firewall server. The data used came directly from logs of the server's login access, updated daily.

A two-way classifier, the initial model received an accuracy of 0.906 with the test datasets, while the final classifier gives predictions with an accuracy of 1.0.